Move all helper methods to a single file
