+++ test/core_test.py
@@ -2,94 +2,7 @@ import unittest
 
 from parse_this import create_parser
 from parse_this.exception import ParseThisException
-
-
-def no_docstring():
-    pass
-
-
-def with_args(a, b):
-    pass
-
-
-def blank_line_in_wrong_place(one: int, two: int):
-    """I put the blank line after arguments ...
-
-    Args:
-        one: this help is #1
-
-        two: this once won't appear sadly
-    """
-    return one * two
-
-
-def parse_me_full_docstring(one: str, two: int, three: int = 12):
-    """Could use some parsing.
-
-    Args:
-        one: some stuff shouldn't be written down
-        two: I can turn 2 syllables words into 6 syllables words
-        three: I don't like the number three
-
-    Returns:
-        the first string argument concatenated with itself 'two' times and the
-        last parameters multiplied by itself
-    """
-    return one * two, three * three
-
-
-def parse_me_no_docstring(one: int, two: int, three: int):
-    return one * two, three * three
-
-
-def multiline_docstring(one: int, two: int, three: int):
-    """I am a sneaky function.
-
-    Args:
-        one: this one is a no brainer
-        three: noticed you're missing docstring for two and
-          I'm multiline too!
-
-    Returns:
-        the first string argument concatenated with itself 'two' times and the
-        last parameters multiplied by itself
-    """
-    return one * two, three * three
-
-
-def different_delimiter_chars(one: int, two: int, three: int):
-    """I am a sneaky function.
-
-    Args:
-        one -- this one is a no brainer even with dashes
-        three -- noticed you're missing docstring for two and
-          I'm multiline too!
-
-    Returns:
-        the first string argument concatenated with itself 'two' times and the
-        last parameters multiplied by itself
-    """
-    return one * two, three * three
-
-
-@create_parser()
-def concatenate_string(string: str, nb_concat: int):
-    return string * nb_concat
-
-
-@create_parser()
-def has_none_default_value(a: int, b: str = None):
-    return a, b
-
-
-@create_parser()
-def has_flags(a: int, b: bool = False):
-    return a, b
-
-
-@create_parser()
-def has_bool_arguments(a: bool):
-    return a
+from test.helpers import has_bool_arguments, has_flags, has_none_default_value
 
 
 class TestCore(unittest.TestCase):
